LayoutTests/fast/canvas/canvas-overloads-strokeRect-expected.txt

Test the behavior of CanvasRenderingContext2D.strokeRect() when called with different numbers of arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.strokeRect() is undefined
PASS ctx.strokeRect(0) is undefined
PASS ctx.strokeRect(0, 0) is undefined
PASS ctx.strokeRect(0, 0, 0) is undefined
PASS ctx.strokeRect(0, 0, 0, 0) is undefined
PASS ctx.strokeRect(0, 0, 0, 0, 0) is undefined
PASS successfullyParsed is true

TEST COMPLETE
